test: 로그인 API 문서 작성
